refactor(simulate): move iter_live_eval_outputs into eval_summary
Fold the shared eval-outputs filter helper into simulate/utils/eval_summary.py
alongside the other free-function eval helpers and delete the standalone
simulate/utils/eval_outputs.py module. Repoint the two import sites in the
test-execution serializer and run_test view. Pure module move; behavior
unchanged. Addresses round-2 follow-up on PR #1147.
